+++ lib/src/pdfium/pdfrx_pdfium.dart
@@ -9,6 +9,7 @@ import 'dart:ui' as ui;
 import 'package:ffi/ffi.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:rxdart/rxdart.dart';
 
 import '../pdf_api.dart';
 import '../utils/unmodifiable_list.dart';
@@ -320,11 +321,15 @@ class _PdfDocumentPdfium extends PdfDocument {
   final pdfium_bindings.FPDF_FORMHANDLE formHandle;
   final Pointer<pdfium_bindings.FPDF_FORMFILLINFO> formInfo;
   bool isDisposed = false;
+  final subject = BehaviorSubject<PdfDocumentEvent>();
 
   @override
   bool get isEncrypted => securityHandlerRevision != -1;
   @override
   final PdfPermissions? permissions;
+
+  @override
+  Stream<PdfDocumentEvent> get events => subject.stream;
 
   _PdfDocumentPdfium._(
     this.document, {
@@ -417,6 +422,9 @@ class _PdfDocumentPdfium extends PdfDocument {
       );
       if (isDisposed) return;
       _pages = List.unmodifiable(loaded.pages);
+
+      subject.add(PdfDocumentPageStatusChangedEvent(this, _pages.sublist(firstUnloadedPageIndex)));
+
       if (onPageLoadProgress != null) {
         final result = await onPageLoadProgress(loaded.pageCountLoadedTotal, loaded.pages.length, data);
         if (result == false) {
@@ -521,6 +529,7 @@ class _PdfDocumentPdfium extends PdfDocument {
   Future<void> dispose() async {
     if (!isDisposed) {
       isDisposed = true;
+      subject.close();
       await (await backgroundWorker).compute((params) {
         final formHandle = pdfium_bindings.FPDF_FORMHANDLE.fromAddress(params.formHandle);
         final formInfo = Pointer<pdfium_bindings.FPDF_FORMFILLINFO>.fromAddress(params.formInfo);
@@ -729,8 +738,11 @@ class _PdfPagePdfium extends PdfPage {
   Future<PdfPageText> loadText() => _PdfPageTextPdfium._loadText(this);
 
   @override
-  Future<List<PdfLink>> loadLinks({bool compact = false}) async {
-    final links = await _loadAnnotLinks() + await _loadLinks();
+  Future<List<PdfLink>> loadLinks({bool compact = false, bool enableAutoLinkDetection = true}) async {
+    final links = await _loadAnnotLinks();
+    if (enableAutoLinkDetection) {
+      links.addAll(await _loadWebLinks());
+    }
     if (compact) {
       for (int i = 0; i < links.length; i++) {
         links[i] = links[i].compact();
@@ -739,7 +751,7 @@ class _PdfPagePdfium extends PdfPage {
     return List.unmodifiable(links);
   }
 
-  Future<List<PdfLink>> _loadLinks() async =>
+  Future<List<PdfLink>> _loadWebLinks() async =>
       document.isDisposed
           ? []
           : await (await backgroundWorker).compute((params) {